refactor(kube): PodIdentity 및 helper를 internal/kube로 승격
netobs/gpuobs가 동일한 식별 어휘를 공유하도록 PodIdentity, IdentityClass 상수, 라벨/키 메서드, TrimGeneratedSuffix 헬퍼를 새 internal/kube 패키지로 옮긴다. netobs/types는 eBPF stage와 EnrichedEvent 표현만 남기고 Src/Dst 필드를 kube.PodIdentity로 교체한다.
